Move GpuPartitioning to a separate file
